refactor(tests): split dashscope provider tests
